refactor user service and split profile image methods into new service
