+++ src/main/java/es/org/cxn/backapp/service/UserService.java
@@ -26,13 +26,9 @@ package es.org.cxn.backapp.service;
 
 import java.util.List;
 
-import org.springframework.web.multipart.MultipartFile;
-
 import es.org.cxn.backapp.exceptions.UserServiceException;
 import es.org.cxn.backapp.model.UserEntity;
 import es.org.cxn.backapp.model.UserRoleName;
-import es.org.cxn.backapp.model.form.responses.ProfileImageResponse;
-import es.org.cxn.backapp.model.persistence.PersistentUserEntity;
 import es.org.cxn.backapp.model.persistence.PersistentUserEntity.UserType;
 import es.org.cxn.backapp.service.dto.UserRegistrationDetailsDto;
 import es.org.cxn.backapp.service.dto.UserServiceUpdateDto;
@@ -143,15 +139,6 @@ public interface UserService {
     List<UserEntity> getAll();
 
     /**
-     * Get user profile image.
-     *
-     * @param dni The user identifier.
-     * @return The profile image response dto.
-     * @throws UserServiceException When user with dni not found.
-     */
-    ProfileImageResponse getProfileImage(String dni) throws UserServiceException;
-
-    /**
      * Removes an user from persistence.
      *
      * @param email email of the user to remove.
@@ -159,36 +146,6 @@ public interface UserService {
      * @throws UserServiceException when user with provided email not found.
      */
     void remove(String email) throws UserServiceException;
-
-    /**
-     * Saves the profile image for a user based on the provided URL.
-     *
-     * @param userDni the DNI (Document Number of Identification) of the user for
-     *                whom the profile image is being saved.
-     * @param url     the URL where the profile image is located. This can be a
-     *                direct URL or a relative path to the image file.
-     * @return the saved {@link PersistentUserEntity} object containing the profile
-     *         image information.
-     * @throws UserServiceException if there is an error while saving the profile
-     *                              image for the user.
-     */
-    PersistentUserEntity saveProfileImage(String userDni, String url) throws UserServiceException;
-
-    /**
-     * Saves the profile image for a user based on the uploaded file.
-     *
-     * @param userDni the DNI (Document Number of Identification) of the user for
-     *                whom the profile image is being saved.
-     * @param file    the uploaded {@link MultipartFile} representing the profile
-     *                image.
-     * @return the saved {@link PersistentUserEntity} object containing the profile
-     *         image information.
-     * @throws IllegalStateException if the method is invoked when the application
-     *                               is in an invalid state for this operation.
-     * @throws UserServiceException  if there is an error while saving the profile
-     *                               image for the user.
-     */
-    PersistentUserEntity saveProfileImageFile(String userDni, MultipartFile file) throws UserServiceException;
 
     /**
      * Unsubscribe an user.
